+++ src/Providers/WAccountServiceProvider.php
@@ -3,7 +3,8 @@
 namespace Werify\Account\Laravel\Providers;
 
 use Illuminate\Support\ServiceProvider;
-use Werify\Account\Laravel\Http\Middleware\V1\Auth;
+use Werify\Account\Laravel\Http\Middleware\V1\WApi;
+use Werify\Account\Laravel\Http\Middleware\V1\WAuth;
 
 class WAccountServiceProvider extends ServiceProvider
 {
@@ -14,7 +15,8 @@ class WAccountServiceProvider extends ServiceProvider
 
     public function boot()
     {
-        app('router')->aliasMiddleware('wauth', Auth::class);
+        app('router')->aliasMiddleware('wauth', WAuth::class);
+        app('router')->aliasMiddleware('wapi', WApi::class);
         if (config('waccount.routes.api.enabled')) {
             $this->loadRoutesFrom(__DIR__.'/../Routes/api.php');
         }