+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 2023-09-04
+
+### Added
+ - Add `WApi` middleware for api usage, You can use `wapi` as middleware, that`s it.
+
+### Changed
+ - Change `WAuth` middleware class name from `Auth.php` to `WAuth.php`.
+
+
 ## [Unreleased] - 2023-08-23
 
 ### Added